+++ packages/web-components/@sfx/products/src/products-search.ts
@@ -1,0 +1,69 @@
+import { css, CSSResult, customElement } from 'lit-element';
+import {
+  SEARCH_RESPONSE,
+  SearchResponsePayload,
+} from '@sfx/events';
+import { ProductsBase } from '.';
+
+/**
+ * The `sfx-products` web component wraps and renders a number of
+ * `sfx-product` components. It wraps each `sfx-product` component in an
+ * additional wrapper for flexibility.
+ *
+ * This component updates upon receiving a [[SEARCH_RESPONSE]] event.
+ */
+@customElement('sfx-products')
+export default class ProductsSearch extends ProductsBase {
+  /**
+   * Binds relevant methods.
+   */
+  constructor() {
+    super();
+
+    this.setProductsFromEvent = this.setProductsFromEvent.bind(this);
+  }
+
+  /**
+   * Registers event listeners.
+   */
+  connectedCallback() {
+    super.connectedCallback();
+
+    window.addEventListener(SEARCH_RESPONSE, this.setProductsFromEvent);
+  }
+
+  /**
+   * Removes event listeners.
+   */
+  disconnectedCallback() {
+    super.disconnectedCallback();
+
+    window.removeEventListener(SEARCH_RESPONSE, this.setProductsFromEvent);
+  }
+
+  /**
+   * Sets the `products` property from the products in an event.
+   *
+   * @param event An event containing a search result with product data.
+   */
+  setProductsFromEvent(event: CustomEvent<SearchResponsePayload>) {
+    const eventGroup = event.detail.group || '';
+    const componentGroup = this.group || '';
+    if (eventGroup === componentGroup) {
+      this.products = event.detail.results.records || [];
+    }
+  }
+
+  protected renderStyles(): CSSResult {
+    return css`
+      sfx-products {
+        display: flex;
+        flex-wrap: wrap;
+      }
+
+      sfx-products[hidden] {
+        display: none;
+      }
+    `;
+  }
+}