+++ packages/web-components/@sfx/products/test/unit/products.test.ts
@@ -1,16 +1,27 @@
 import { expect, sinon, stub } from '../utils';
-import Products from '../../src/products';
-import { SAYT_PRODUCTS_RESPONSE } from '@sfx/events';
+import { SAYT_PRODUCTS_RESPONSE, SEARCH_RESPONSE } from '@sfx/events';
+import ProductsBase from '../../src/products-base';
+import ProductsSayt from '../../src/products-sayt';
+import ProductsSearch from '../../src/products-search';
 
-describe('Products Component', () => {
+describe('Products Base Component', () => {
   let component;
+
   beforeEach(() => {
-    component = new Products();
+    component = new ProductsBase();
   });
 
   describe('constructor', () => {
-    it('should default to have an empty array of products', () => {
-      expect(component.products).to.deep.equal([]);
+    describe('products property', () => {
+      it('should default to be an empty array of products', () => {
+        expect(component.products).to.deep.equal([]);
+      });
+    });
+
+    describe('group property', () => {
+      it('should have default value of an empty string', () => {
+        expect(component.group).to.equal('');
+      });
     });
   });
 
@@ -40,8 +51,26 @@ describe('Products Component', () => {
 
       expect(setAttribute).to.not.be.called;
     });
+  });
+});
 
-    it('should set up event listener for a provide-products event to set products', () => {
+describe('Products Sayt Component', () => {
+  let component;
+
+  beforeEach(() => {
+    component = new ProductsSayt();
+  });
+
+  describe('connectedCallback', () => {
+    it('should call super', () => {
+      const superConnected = stub(Object.getPrototypeOf(component), 'connectedCallback');
+
+      component.connectedCallback();
+
+      expect(superConnected).to.be.calledOnce;
+    });
+
+    it('should set up an event listener for a sayt products event to set products', () => {
       const addEventListener = sinon.stub(window, 'addEventListener');
 
       component.connectedCallback();
@@ -59,7 +88,7 @@ describe('Products Component', () => {
       expect(superDisconnected).to.be.calledOnce;
     });
 
-    it('should remove provide-products event listener', () => {
+    it('should remove an event listener for sayt products', () => {
       const removeEventListener = sinon.stub(window, 'removeEventListener');
 
       component.disconnectedCallback();
@@ -69,8 +98,39 @@ describe('Products Component', () => {
   });
 
   describe('setProductsFromEvent', () => {
-    it('should set the event products payload into the component', () => {
-      const products = [1, 2, 3];
+    let products;
+    let group;
+
+    beforeEach(() => {
+      products = [1, 2, 3];
+      group = 'group';
+    });
+
+    it('should set products to an empty array if the event payload does not contain products', () => {
+      const event = { detail: {} };
+      component.setProductsFromEvent(event);
+
+      expect(component.products).to.deep.equal([]);
+    });
+
+    it('should set products when the event matches the group in the component', () => {
+      const event = { detail: { products, group } };
+      component.group = group;
+
+      component.setProductsFromEvent(event);
+
+      expect(component.products).to.equal(products);
+    });
+
+    it('should not set products when the group in the component and event do not match', () => {
+      const event = { detail: { products, group } };
+
+      component.setProductsFromEvent(event);
+
+      expect(component.products).to.deep.equal([]);
+    });
+
+    it('should default the group in the event to an empty string if it is falsey', () => {
       const event = { detail: { products } };
 
       component.setProductsFromEvent(event);
@@ -78,12 +138,108 @@ describe('Products Component', () => {
       expect(component.products).to.equal(products);
     });
 
-    it('should set the products property to an empty array if payload of the event is undefined', () => {
+    it('should default the group in the component to an empty string if it is falsey', () => {
+      component.group = undefined;
+      const event = { detail: { products } };
+
+      component.setProductsFromEvent(event);
+
+      expect(component.products).to.equal(products);
+    });
+  });
+});
+
+describe('Products Search Component', () => {
+  let component;
+
+  beforeEach(() => {
+    component = new ProductsSearch();
+  });
+
+  describe('connectedCallback', () => {
+    it('should call super', () => {
+      const superConnected = stub(Object.getPrototypeOf(component), 'connectedCallback');
+
+      component.connectedCallback();
+
+      expect(superConnected).to.be.calledOnce;
+    });
+
+    it('should set up an event listener for a search response to set products', () => {
+      const addEventListener = sinon.stub(window, 'addEventListener');
+
+      component.connectedCallback();
+
+      expect(addEventListener).to.be.calledWith(SEARCH_RESPONSE, component.setProductsFromEvent);
+    });
+  });
+
+  describe('disconnectedCallback', () => {
+    it('should call super', () => {
+      const superDisconnected = stub(Object.getPrototypeOf(component), 'disconnectedCallback');
+
+      component.disconnectedCallback();
+
+      expect(superDisconnected).to.be.calledOnce;
+    });
+
+    it('should remove an event listener for search responses', () => {
+      const removeEventListener = sinon.stub(window, 'removeEventListener');
+
+      component.disconnectedCallback();
+
+      expect(removeEventListener).to.be.calledWith(SEARCH_RESPONSE, component.setProductsFromEvent);
+    });
+  });
+
+  describe('setProductsFromEvent', () => {
+    let records;
+    let group;
+
+    beforeEach(() => {
+      records = [1, 2, 3];
+      group = 'group';
+    });
+
+    it('should set products to an empty array if the event payload does not contain records', () => {
       const event = { detail: { results: {} } };
+      component.setProductsFromEvent(event);
+
+      expect(component.products).to.deep.equal([]);
+    });
+
+    it('should set products when the event matches the group in the component', () => {
+      const event = { detail: { results: { records }, group } };
+      component.group = group;
+
+      component.setProductsFromEvent(event);
+
+      expect(component.products).to.equal(records);
+    });
+
+    it('should not set products when the group in the component and event do not match', () => {
+      const event = { detail: { results: { records }, group } };
 
       component.setProductsFromEvent(event);
 
       expect(component.products).to.deep.equal([]);
+    });
+
+    it('should default the group in the event to an empty string if it is falsey', () => {
+      const event = { detail: { results: { records } } };
+
+      component.setProductsFromEvent(event);
+
+      expect(component.products).to.equal(records);
+    });
+
+    it('should default the group in the component to an empty string if it is falsey', () => {
+      component.group = undefined;
+      const event = { detail: { results: { records } } };
+
+      component.setProductsFromEvent(event);
+
+      expect(component.products).to.equal(records);
     });
   });
 });